Fix zero-divide; avoid histogram overflow; refactor ThreadDump
Root cause:\n- Maths.divideRoundUp could divide by zero when divisor==0 (javabugs:S3518).\n- RecordingHistogram computed offset via long subtraction before conversion, risking overflow (java:S2184).\n- Some tests lacked explicit assertions (java:S2699).\n- Empty bootstrap method flagged as empty (java:S1186).\n- Wildcard generics in enum lookup caused unsafe types (java:S1452).\n- ThreadDump.assertNoNewThreads had high cognitive complexity (java:S3776).\n\nFix:\n- Throw IllegalArgumentException for divisor==0 and document the contract; add a unit test.\n- Perform offset subtraction in floating-point before conversion to avoid overflow.\n- Add assertDoesNotThrow to ExceptionHandlerFallbackTest; adjust ObjectUtils tests to strong generics.\n- Document intentional no-op in ChronicleGuarding.bootstrap().\n- Make caseIgnoreLookup generic and adapt ClassLocal initialiser with a narrow, documented cast.\n- Extract helper to reduce branching in ThreadDump.assertNoNewThreads.\n\nImpact:\n- Prevents arithmetic error and offset overflow; clearer behavioural contract.\n- Reduces cognitive complexity and improves readability without API breakage.\n- Strengthens type-safety for enum lookups; targeted tests stay green offline (mvn -o).

+++ src/main/java/net/openhft/chronicle/core/util/RecordingHistogram.java
@@ -160,7 +160,8 @@ public class RecordingHistogram extends Histogram {
             sb.append("[");
             String sep = "";
             for (int i = 0, lim = Math.min(Math.min(20, max * 2), count * 2); i < lim; i += 2) {
-                double offset = toMicros.apply(top[i] - start);
+                // ensure subtraction is performed in floating-point to avoid long overflow prior to conversion
+                double offset = toMicros.apply(((double) top[i]) - ((double) start));
                 double duration = toMicros.apply(top[i + 1]);
                 sb.append(sep).append("{ off: ").append(offset).append(", dur: ").append(duration).append(" }");
                 sep = ", ";
